Do no show attack restart point above candidates count
It is possible that parallel workers explore all remaining candidates
before the user's interruption or early exit actually occurs. This can
happen for example if the solution is close to the end of the list of
candidates. Then it does not make sense to restart the attack so a
restart point is not needed.